requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.1.0
bleach==6.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coinbase-commerce==1.0.1
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
daphne==4.0.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.2
django-cors-headers==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-extra-fields==3.5.0
drf-yasg==1.21.6
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
filetype==1.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
ipython==8.12.3
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==23.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.18.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==26.0.3
redis==4.6.0
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.0.3
user-agents==2.2.0
vine==5.0.0
virtualenv==20.26.2
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
yarg==0.1.9
zope.interface==6.0